requirements.txt
Django==3.1
djangorestframework==3.11.1
pandas==1.1.0